Improved Serial and Processing3 Display
Updated the serial data format to better support the Processing3 OScope display. Four channels of 512 bytes are transfered each time the local buffer fills up.
Development is finish for this version. The Rev3 board is cleaner and should be noise free (to the degree that I can make it) and easier to interface to since it also has build in USB for power and serial.